requirements.txt
# Requirements pour SimpleBooth - Raspberry Pi OS
# Versions compatibles Python 3.13 et ARM64
# Dernière mise à jour: 2025-11-09

# === CORE DEPENDENCIES ===
# NumPy - Version compatible Python 3.13
numpy>=1.26.0

# OpenCV - Version compatible Python 3.13 (sans version fixe pour wheels)
opencv-python>=4.8.0

# === WEB FRAMEWORK ===
# Flask et dépendances
Flask>=2.3.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
click>=8.1.0
itsdangerous>=2.1.0

# === IMAGE PROCESSING ===
# Pillow - Traitement d'images
Pillow>=10.0.0

# === HARDWARE INTERFACES ===
# Communication série pour imprimante
pyserial>=3.5

# Impression thermique
python-escpos>=3.0

# Caméra Raspberry Pi (si utilisée)
picamera2>=0.3.0

# === COMMUNICATION ===
# Telegram Bot (optionnel)
python-telegram-bot>=20.0

# Requêtes HTTP
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.0.0
charset-normalizer>=3.3.0
idna>=3.4

# === SYSTEM UTILITIES ===
# Gestion des processus
psutil>=5.9.0

# Configuration et logging
PyYAML>=6.0.0
coloredlogs>=15.0.0
humanfriendly>=10.0
